factor out common attrs
